Import py27-yapf-0.1.6 as wip/py-yapf.
YAPF - Yet Another Python Formatter - is a python formatting
tool based based off of 'clang-format'.
In essence, the algorithm takes the code and reformats it to
the best formatting that conforms to the style guide, even if
the original code didn't violate the style guide.
The idea is also similar to the 'gofmt' tool for the Go
programming language: end all holy wars about formatting -
if the whole code base of a project is simply piped through
YAPF whenever modifications are made, the style remains consistent
throughout the project and there's no point arguing about style
in every code review.