Import libunwind-1.1 as wip/libunwind
This package is just a skeleton of libunwind, not supported on NetBSD.
The support is work-in-progress upstream.

Libunwind defines a portable and efficient C programming interface (API) to
determine the call-chain of a program. The API additionally provides the means
to manipulate the preserved (callee-saved) state of each call-frame and to
resume execution at any point in the call-chain (non-local goto). The API
supports both local (same-process) and remote (across-process) operation.