Import meta-pkgs/gnome as wip/gnome and bump version to 3.17
gnome3 packages are required directly or indirectly by other DE, like
MATE or LXQt, so it's better to get it finally there.

DESCR:

GNOME is the GNU Network Object Model Environment.

The GNOME project intends to build a complete, easy-to-use desktop
environment for the user, and a powerful application framework for
the software developer.

This is a meta package for the second major version of the GNOME
project.  It provides the main GNOME applications, defined by the
project as the "GNOME desktop".