Import pristine x11/zenity-2.32.1nb19 as wip/zenity
This package will be upgraded for Gnome 3

This is zenity - a rewrite of gdialog, the GNOME port of dialog
which allows you to display dialog boxes from the command line
and shell scripts. If you understand, things are just as they
are. If you don't understand, things are just as they are.