Import py27-Kallithea-0.2.2 as wip/py-Kallithea.
Kallithea is a GPLv3'd, Free Software source code management system
that supports two leading version control systems, Mercurial and
Git, and has a web interface that is easy to use for users and
admins. You can install Kallithea on your own server and host
repositories for the version control system of your choice.

TODO: test if it works.
All dependencies should be there, but some upper bounds are disregarded
because pkgsrc has newer versions and I didn't want to import old versions
just for testing this package; we can if we find out they're needed.