Update wip/chicken to chicken-4.10.0rc4.
Changes:
The changes since 4.10.0rc1 are as follows:
- SIGBUS, SIGILL and SIGFPE cause proper exceptions instead of panicking
- string-translate* no longer scans beyond string boundaries (CVE-2015-4556)
- A corner case bug in substring-index and substring-index-ci that was
   introduced as part of the fix for CVE-2014-9651 has been fixed.
- chicken.h has been fixed so it does not cause an error in C++ mode.
- create-directory will not skip creation of topmost parent dir in Windows.
- The types.db entries for {read,write}-string read-u8vector have been improved.
- The types.db entries for several SRFI-4 procedures have been fixed.
- The performance of read-u8vector has been improved.
- The "scheme" module now includes an export for the "angle" procedure.
- normalize-pathname no longer considers ".//foo" as the absolute "/foo".
- A few small improvements have been made in the manual.

If you're wondering what happened to release candidate 2 and 3, the
rc2 was created but never announced, and rc3 was incorrectly versioned,
so it has been skipped.  Sorry for that, as well :)

Please test this new version on as many platforms as you can, and let
us know if there are any problems with it.  If no problems are found
we can make a 4.10.0 release soon.