Update to elinks-current-0.13
This snapshot package was initially created to provide version 0.12pre6 when
www/elinks was on 0.11.3. However, www/elinks has been on 0.12pre6 for some
time now, making this wip package redundant.

PR 50105 was created to update www/elinks to elinks-current-0.13 because the
0.12pre6 version would fail to render some sites due to SSL connection errors.
This problem was reported on the ELinks Users mailing list and the advice was
to use version 0.13 which fixes this issue. After review and discussion it was
realised that 0.12pre6 is still the most recent/current unstable version,
however. No 0.13 tag exists in the Elinks git tree because Elinks is still
being developed against this version number. Therefore elinks-current-0.13 is a
development snapshot, albeit one that does not change frequently.

Therefore it was decided to close PR 50105 and update this snapshot package to
provide the 0.13 version.