Import cataclysm-dda-0.12 as wip/cataclysm-dda.
Cataclysm: Dark Days Ahead is a roguelike set in a post-apocalyptic
world. While some have described it as a "zombie game", there's
far more to Cataclysm than that. Struggle to survive in a harsh,
persistent, procedurally generated world. Scavenge the remnants of
a dead civilization for for food, equipment, or, if you're lucky,
a vehicle with a full tank of gas to get you the hell out of Dodge.
Fight to defeat or escape from a wide variety of powerful monstrosities,
from zombies to giant insects to killer robots and things far
stranger and deadlier, and against the others like yourself, that
want what you have...