Import R-Cairo-1.5.8 as wip/R-cairo.
This package provides a Cairo graphics device that can be use to
create high-quality vector (PDF, PostScript and SVG) and bitmap
output (PNG,JPEG,TIFF), and high-quality rendering in displays (X11
and Win32). Since it uses the same back-end for all output, copying
across formats is WYSIWYG. Files are created without the dependence
on X11 or other external programs.  This device supports alpha
channel (semi-transparent drawing) and resulting images can contain
transparent and semi-transparent regions. It is ideal for use in
server environments (file output) and as a replacement for other
devices that don't have Cairo's capabilities such as alpha support
or anti-aliasing. Backends are modular such that any subset of
backends is supported.