Import infraarcana-17.0 as wip/infraarcana.
Infra: below, underneath, in the underworld
Arcana: secrets, mystery, hidden knowledge

Infra Arcana is a Roguelike set in the early 20th century. The goal
is to explore the lair of a dreaded cult called The Church of Starry
Wisdom. Buried deep beneath their hallowed grounds lies an artifact
called The Shining Trapezohedron - a window to all secrets of the
universe. Your ultimate goal is to unearth this artifact.