Import xoreos-0.0.2 as wip/xoreos.
xoreos is an open source implementation of BioWare's Aurora engine and its
derivatives, licensed under the terms of the [GNU General Public License
version 3](https://www.gnu.org/licenses/gpl.html) (or later). The goal is
to have all games using this engines working in a portable manner, starting
from Neverwinter Nights and ending with Dragon Age II.

This is a game engine - the data files are not included.