Import Commander-Genius-1822 as wip/commandergenius.
Commander-Genius aka CloneKeenPlus is an engine which interprets
Commander Keen 1-6 (Vorticons and Galaxy Series).

As fans and developers we try to implement new features, improve
the gameplay and give the feeling, you are playing the original
game.

Obviously you need the game data of the games to play it.  Commander
Keen 1 and 4 come included in this package.  Episode 2, 3, 5 and
6 are registered versions which can be purchased by 3D Realms or
Apogee, or bought by someone else.  If you still have the old games
of the registered version, then give it a try. You will be surprised,
how well they will look!