Import gts.121130 as wip/gts-snapshot.
This is the GTS library (a version called snaphot by the authors).
GTS stands for the GNU Triangulated Surface Library. It includes a
number of useful functions to deal with triangulated surfaces
including, but not limited to, multi-resolution models, Delaunay and
Constrained Delaunay triangulations, set operations on surfaces
(intersection, union etc ...), bounding-boxes trees for efficient
collision and intersection detection, triangle strips generation for
fast rendering.