Import virtualbox-ose as virtualbox-ose-4.3.30  - This wip version is purely experimental, it is on the way, not completed     at all, or just started to compile by wip/kBuild-svn     kBuild-0.1.5nb20150905 Build framework based on GNU make
 - This is ported from FreeBSD/ports (4.3.30)
    See http://svnweb.freebsd.org/ports/head/emulators/virtualbox-ose/
 - Thus there should be so many things to adjust from FreeBSD to pkgsrc

VirtualBox is a family of powerful x86 virtualization products for
enterprise as well as home use. Not only is VirtualBox an extremely
feature rich, high performance product for enterprise customers, it
is also the only professional solution that is freely available as
Open Source Software under the terms of the GNU General Public License.

WWW: http://www.virtualbox.org/