Import openh264-1.4.0 as wip/openh264.
Partly based on the FreeBSD ports package. Goes into wip because firefox
is not loading the plugin properly for some reason.

OpenH264 is a codec library which supports H.264 encoding and decoding.
It is suitable for use in real time applications such as WebRTC.
This package also contains the OpenH264 plugin for www/firefox.