angelscript: import wip/angelscript as angelscript-2.30.2
The AngelCode Scripting Library, or AngelScript as it is also known, is an
extremely flexible cross-platform scripting library designed to allow
applications to extend their functionality through external scripts. It has
been designed from the beginning to be an easy to use component, both for
the application programmer and the script writer.

Efforts have been made to let it call standard C functions and C++ methods
with little to no need for proxy functions. The application simply registers
the functions, objects, and methods that the scripts should be able to work
with and nothing more has to be done with your code. The same functions used
by the application internally can also be used by the scripting engine,
which eliminates the need to duplicate functionality.