Update Ledger to the latest release 3.1
Note that 3.1 is still a year old and 3.1.1 must be just around the
corner from being released, but this package had become so out of date
that the urls were no longer correct. I thought targeting a released
version was a better idea than a git branch, etc.

I have built and tested this on NetBSD 7 AMD64 and it works, but I have
had to use a bit of a hacky approach in the Makefile. Any tips on a
cleaner approach would be appreciated. See the Makefile and the TODO
file for more information on what I've tried.