Import lumina-0.8.7 as wip/lumina.
The Lumina Desktop Environment is a lightweight system interface that is
designed for use on any Unix-like operating system. It takes a plugin-based
approach, allowing the entire interface to be assembled/arranged by each
individual user as desired, with a system-wide default layout which was setup
by the  system administrator. This allows every system (or user session) to be
designed to maximize the individual s productivity.

The Lumina desktop developers understand that the point of a computer system is
to run applications, so Lumina was designed to require as few system
dependencies/requirements as possible. This allows it to be used to revitalize
older systems or to allow the user to run applications that may need a higher
percentage of the system resources than were previously available with other
desktop environments.

All of this results in a very lightweight, customizable, and smooth desktop
experience with minimal system overhead.