Import sonic-visualiser-2.5 as wip/sonic-visualiser.
Sonic Visualiser is an application for viewing and analysing the contents of
music audio files.

The aim of Sonic Visualiser is to be the first program you reach for when want
to study a musical recording rather than simply listen to it.

We hope Sonic Visualiser will be of particular interest to musicologists,
archivists, signal-processing researchers and anyone else looking for a friendly
way to take a look at what lies inside the audio file.

It was developed at the Centre for Digital Music at Queen Mary, University of
London.