Add libfpx.
Can be used by Graphicsmagick to read flashpix images, not in pkgsrc as mention
by richard palo on pkgsrc-users.

Libfpx is a library for reading and writing FlashPix images.

This product includes software developed by the contributors and Digital
Imaging Group, Inc. (http://www.digitalimaging.org/) for use in the
Flashpix Toolkit Project.