Add an unfinished port for include-what-you-use.
"Include what you use" means this: for every symbol (type, function
variable, or macro) that you use in foo.cc, either foo.cc or foo.h
should #include a .h file that exports the declaration of that symbol.
The include-what-you-use tool is a program that can be built with the
clang libraries in order to analyze #includes of source files to find
include-what-you-use violations, and suggest fixes for them.

The main goal of include-what-you-use is to remove superfluous #includes.
It does this both by figuring out what #includes are not actually needed
for this file (for both .cc and .h files), and replacing #includes with
forward-declares when possible.