edb-debugger: message($$QMAKE_HOST.arch) shows amd64 on NetBSD
This looks odd, we seem to be the only known to me platform which uses
QMAKE_HOST.arch=amd64, instead of x86_64.

OpenBSD, FreeBSD, Linux, Windows and MacOSX define QMAKE_HOST.arch as x86_64.

According to:
https://github.com/malikcjm/qtcreator/blob/cd7b6e0414c867f708ada0d1cbb4c0173a27be4b/src/shared/proparser/qmakeevaluator.cpp#L1034

QMAKE_HOST.arch maps to `sysctl hw.machine` and it's represented as
'amd64' on NetBSD. So it looks like we are unique here.