Add disque-1.0-rc1 as wip/disque.
Disque is an ongoing experiment to build a distributed, in-memory, message
broker. Its goal is to capture the essence of the "Redis as a jobs queue"
use case, which is usually implemented using blocking list operations, and
move it into an ad-hoc, self-contained, scalable, and fault tolerant design,
with simple to understand properties and guarantees, but still resembling
Redis in terms of simplicity, performance, and implementation as a C
non-blocking networked server.