Import logorate-3.9.2 as wip/logrotate-git.
The logrotate utility is designed to simplify the administration of
log files on a system which generates a lot of log files.  Logrotate
allows for the automatic rotation compression, removal and mailing of
log files.  Logrotate can be set to handle a log file daily, weekly,
monthly or when the log file gets to a certain size.  Normally,
logrotate runs as a daily cron job.

This package will replace wip/logrotate-svn since the project is now
hosted on GitHub.