pioneer: Import pioneer-20160116 as wip/pioneer.
Pioneer is a space adventure game set in the milkyway galaxy at the turn of
the 31st century. The game is open-ended, and you are free to explore the
millions of star systems in the game. You can land on planets, slingshot
past gas giants, and burn yourself to a crisp flying between binary star
systems.