hg-package.mk: avoid cache extraction if hg repo exists in the work dir
When working with huge hg packages (like firefox) the caching and cache
extraction of the hg repository and the working copy take a lot of time.

When we find a .hg directory in the work dir assume it is a full hg
repository clone, leave it intact, pull it to current and checkout
(throwing away local modifications).