xoreos: update to xoreos-0.0.4
Summary:

Neverwinter Nights now shows speech bubbles for conversation
one-liners, as used for cutscenes, bark strings and short NPC
dialogues. Additionally, the premium modules BioWare sold for
Neverwinter Nights, including the three that come with the Diamond
Edition, can now be properly loaded and started.

An oversight in the handling of the texture fonts used in Neverwinter
Nights and the two Knights of Old Republic games has been fixed.
This oversight broke rendering of certain characters, most prominently
of those used in eastern European languages and the "smart" single
quotation mark that's used instead of an apostrophe in some strings
found in the French versions.

Moreover, this release includes a lot of user-invisible code
documentation and quality fixes.

Please note that xoreos is still missing a GUI and needs to be
started from the command line.

Changes:

 - DOCS:
  - Improved documentation in various files
  - Rewrote manpage to use semantic -mdoc macros
 - BUILD:
  - Fixed various little problems in the CMake system related to
    non-MSVC compilers
  - Added configure/CMake option to let the user select between
    internal and external GLEW libraries
  - Added dynamic supported warning flag detection to CMake
 - GRAPHICS:
  - Made TextureFont codepage-aware
 - ENGINES:
  - NWN:
   - Implemented support for premium modules
   - Implemented speech bubbles
  - DRAGONAGE:
   - Added engine script functions from the 1.01 and 1.03
     updates
 - MISC:
  - Changed C-style casts to static_cast and reinterpret_cast
    in various places
  - Improved const-correctness in various places
  - Ran many file format loaders through afl and fixed all crashes